% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.na.R
\name{data.frame.na}
\alias{data.frame.na}
\title{data.frames with filled NA's}
\usage{
data.frame.na(
  ...,
  row.names = NULL,
  check.rows = FALSE,
  check.names = TRUE,
  stringsAsFactors = FALSE
)
}
\description{
data.frames with filled NA's
}
\keyword{internal}
