% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example1.TS.R
\docType{data}
\name{Example1.TS}
\alias{Example1.TS}
\title{A vector containing a time series}
\format{A vector of length 24}
\usage{
data(Example1.TS)
}
\description{
A vector of turnover values for an arbitrary undisclosed unit of the survey 
IASS
}
