% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example2.TS.R
\docType{data}
\name{Example2.TS}
\alias{Example2.TS}
\title{A vector containing a time series}
\format{A vector of length 24}
\usage{
data(Example2.TS)
}
\description{
A vector of turnover values for an arbitrary undisclosed unit of the survey 
IASS
}
