% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedTSPred.R
\docType{methods}
\name{FixedTSPred}
\alias{FixedTSPred}
\alias{FixedTSPred,vector-method}
\alias{FixedTSPred,StQList-method}
\title{Method to predict a prefixed value for time series.}
\usage{
FixedTSPred(x, VarNames, value = 0L)

\S4method{FixedTSPred}{vector}(x, VarNames, value = 0L)

\S4method{FixedTSPred}{StQList}(x, VarNames, value = 0L)
}
\arguments{
\item{x}{object upon which the prediction will be made.}

\item{VarNames}{character vector with the variable names for which the
prediction will be made; by default it is NULL.}

\item{value}{numeric vector with the value for the prediction and the
sta}
}
\value{
It returns a list with components Pred and STD, containing the point
prediction and the standard deviations, respectively, with the prefixed value.
Depending on the class of the input parameter x, it returns:

\itemize{
 \item For input class vector, it returns numeric vectors.
 \item For input class matrix, it returns matrices.
 \item For input class StQList, it returns list whose components are
  data.tables.
}
}
\description{
This method implements the predicted value and their standard
deviations as a prefixed value.
}
\examples{

# Predicting one and two months ahead in time
data(Example1.TS)
FixedTSPred(Example1.TS, value = 0L)

# Predicting upon a times series with many NA values
data(Example2.TS)
FixedTSPred(Example2.TS, value = 0L)

\dontrun{
# With an object of class StQList
data(StQListExample)
VarNames <- c('ActivEcono_35._6._2.1.4._0', 'GeoLoc_35._6._2.1._1.2.5.')
FixedTSPred(StQListExample, VarNames)
}

}
