% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluating_subclasses.r
\name{MSE_eval}
\alias{MSE_eval}
\title{Prediction/modeling quality metrics}
\usage{
MSE_eval()
}
\value{
An object of class \code{evaluating}.
}
\description{
Constructors for the \code{evaluating} class representing a time series prediction
or modeling fitness quality evaluation based on particular metrics.
}
\section{Error metrics}{

	Mean Squared Error.
}

\seealso{
Other constructors: 
\code{\link{ARIMA}()},
\code{\link{LT}()},
\code{\link{evaluating}()},
\code{\link{modeling}()},
\code{\link{processing}()},
\code{\link{tspred}()}
}
\author{
Rebecca Pontes Salles
}
\concept{constructors}
\keyword{evaluation}
\keyword{metric}
\keyword{quality}
