% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tspred.r
\name{postprocess.tspred}
\alias{postprocess.tspred}
\title{Postprocess method for \code{\link{tspred}} objects}
\usage{
\method{postprocess}{tspred}(obj, ...)
}
\arguments{
\item{obj}{An object of class \code{\link{tspred}} defining a particular time series prediction process.}

\item{...}{Other parameters passed to the method \code{\link{postprocess}} of 
the \code{\link{processing}} objects from \code{obj}.}
}
\value{
An object of class \code{tspred} with updated structure containing
postprocessed time series predictions.
}
\description{
Performs postprocessing of the predicted time series data contained in a \code{\link{tspred}} class object
reversing a particular set of transformation methods. Each transformation method is defined
by a \code{\link{processing}} object in the list contained in the \code{\link{tspred}} class object.
}
\details{
The function \code{\link{postprocess.tspred}} recursively calls the method \code{\link{postprocess}} 
on each \code{\link{processing}} object contained in \code{obj} in the inverse order 
as done by \code{\link{preprocess.tspred}}. The postprocessed predictions 
resulting from each of these calls is used as input to the next call. Finally, the produced 
postprocessed time series predictions are introduced in the structure of the \code{\link{tspred}} class object in \code{obj}.

The same transformation method parameters used/computed during preprocessing, duly saved
in the structure of the \code{\link{tspred}} class object in \code{obj}, are used for 
reversing the transformations during postprocessing.
}
\examples{
\donttest{
data(CATS)

#Obtaining objects of the processing class
proc1 <- subsetting(test_len=20)
proc2 <- BoxCoxT(lambda=NULL)
proc3 <- WT(level=1, filter="bl14")

#Obtaining objects of the modeling class
modl1 <- ARIMA()

#Obtaining objects of the evaluating class
eval1 <- MSE_eval()

#Defining a time series prediction process
tspred_1 <- tspred(subsetting=proc1,
                   processing=list(BCT=proc2,
                                   WT=proc3), 
                   modeling=modl1,
                   evaluating=list(MSE=eval1)
)
summary(tspred_1)

tspred_1 <- subset(tspred_1, data=CATS[3])
tspred_1 <- preprocess(tspred_1,prep_test=FALSE)
tspred_1 <- train(tspred_1)
tspred_1 <- predict(tspred_1, onestep=TRUE)
tspred_1 <- postprocess(tspred_1)
}

}
\seealso{
[tspred()] for defining a particular time series prediction process,
and [LT()] for defining a time series transformation method.

Other preprocess: 
\code{\link{preprocess.tspred}()},
\code{\link{subset}()}
}
\author{
Rebecca Pontes Salles
}
\concept{preprocess}
\keyword{postprocessing}
\keyword{preprocessing}
\keyword{processing}
\keyword{transformation}
