% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SantaFe.D.cont.r
\docType{data}
\name{SantaFe.D.cont}
\alias{SantaFe.D.cont}
\title{Continuation dataset of the time series D of the Santa Fe Time Series
Competition}
\format{
A data frame with 500 observations on the following variable.
\describe{ \item{V1}{a numeric vector containing further
observations of the univariate time series D of the Santa Fe Time Series
Competition in \code{\link{SantaFe.D}}.} }
}
\usage{
SantaFe.D.cont
}
\description{
A univariate time series providing 500 points beyond the end of the time
series D in \code{\link{SantaFe.D}}.
}
\details{
Contains the 500 observations which were to be predicted of the time series
D (\code{\link{SantaFe.D}}) as demanded by the Santa Fe Time Series
Competition.
}
\examples{

data(SantaFe.D.cont)
str(SantaFe.D.cont)
plot(ts(SantaFe.D.cont))

}
\references{
A.S. Weigend, 1993, Time Series Prediction: Forecasting The
Future And Understanding The Past. Reading, MA, Westview Press.
}
\seealso{
\code{\link{SantaFe.D}}, \code{\link{SantaFe.A}},
\code{\link{SantaFe.A.cont}} ~
}
\keyword{Competition}
\keyword{Fe}
\keyword{Santa}
\keyword{Series}
\keyword{Time}
\keyword{datasets}
