% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatDiffTSPred.R
\docType{methods}
\name{StatDiffTSPred}
\alias{StatDiffTSPred}
\alias{StatDiffTSPred,StQList-method}
\alias{StatDiffTSPred,matrix-method}
\alias{StatDiffTSPred,vector-method}
\title{Method to predict according to the stational difference time series
model.}
\usage{
StatDiffTSPred(x, StatDiff = 12L, forward = 2L, VarNames = NULL)

\S4method{StatDiffTSPred}{vector}(x, StatDiff = 12L, forward = 2L,
  VarNames = NULL)

\S4method{StatDiffTSPred}{matrix}(x, StatDiff = 12L, forward = 2L,
  VarNames = NULL)

\S4method{StatDiffTSPred}{StQList}(x, StatDiff = 12L, forward = 2L,
  VarNames = NULL)
}
\arguments{
\item{x}{object upon which the prediction will be made.}

\item{StatDiff}{stational differences of the time series; by default it is
12L.}

\item{forward}{integer indicating the number of periods ahead when the
prediction will be made; by default it is 2L.}

\item{VarNames}{character vector with the variable names for which the
prediction will be made; by default it is NULL}
}
\value{
It returns a list with components Pred and STD, containing the point
prediction and the estimated standard deviations, respectively. Depending
on the class of the input parameter x, it returns:

\itemize{
 \item For input class vector, it returns numeric vectors.
 \item For input class matrix, it returns matrices.
 \item For input class StQList, it returns list whose components are
  data.tables.
}
}
\description{
This method implements the predicted value and their standard
deviation according to the stational difference time series model
\eqn{(1-B)^s (1-B)y_{t}=a_{t}}{(1-B)^s (1-B)y<sub>t</sub>=a<sub>t</sub>}.
}
\examples{

# Predicting one and two months ahead in time
data(Example1.TS)
StatDiffTSPred(Example1.TS, forward = 1L)
StatDiffTSPred(Example1.TS, forward = 2L)

# Predicting upon a times series with many NA values
data(Example2.TS)
StatDiffTSPred(Example2.TS, forward = 1L)

# On a matrix
Mat <- rbind(Example1.TS, Example2.TS)
StatDiffTSPred(Mat, forward = 1L)

# With an object of class StQList
data(StQList_Example)
VarNames <- c('ActivEcono_35._6._2.1.4._0', 'GeoLoc_35._6._2.1._1.2.5.')
StatDiffTSPred(StQList_Example, VarNames = VarNames)

}

