% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StatRegDiffTSPred.R
\docType{methods}
\name{StatRegDiffTSPred}
\alias{StatRegDiffTSPred}
\alias{StatRegDiffTSPred,list-method}
\alias{StatRegDiffTSPred,matrix-method}
\alias{StatRegDiffTSPred,vector-method}
\title{Method to predict according to the stational difference time series model}
\usage{
StatRegDiffTSPred(x, StatDiff = 12L, forward = 2L, VarNames = NULL,
  keyVar = NULL)

\S4method{StatRegDiffTSPred}{vector}(x, StatDiff = 12L, forward = 2L,
  VarNames = NULL, keyVar = NULL)

\S4method{StatRegDiffTSPred}{matrix}(x, StatDiff = 12L, forward = 2L,
  VarNames = NULL, keyVar = NULL)

\S4method{StatRegDiffTSPred}{list}(x, StatDiff = 12L, forward = 2L,
  VarNames = NULL, keyVar = NULL)
}
\arguments{
\item{x}{object upon which the prediction will be made}

\item{StatDiff}{stational difference s of the time series; by default it is 12L}

\item{forward}{integer indicating the number of periods ahead when the
prediction will be made; by default it is 2L}

\item{VarNames}{character vector with the variable names for which the
prediction will be made; by default it is NULL}

\item{keyVar}{character vector with the variable names used as key variables
in data.table operations; by default it is NULL}
}
\value{
It returns a list with components Pred and STD, containing the point
prediction and the estimated standard deviations, respectively. Depending
on the class of the input parameter x, it returns:

\itemize{
 \item For input class vector, it returns numeric vectors
 \item For input class matrix, it returns matrices
 \item For input class data.table, it returns data.tables
 \item For input class StBusQ, it returns data.tables
 \item For input class GenStBusQ, it returns data.tables
}
}
\description{
This method implements the predicted value and their standard deviation
according to the regular difference time series model
\eqn{(1-B)^s y_{t}=a_{t}}{(1-B)^s y<sub>t</sub>=a<sub>t</sub>}
}
\examples{
# Predicting one and two months ahead in time
data(Example1.TS)
StatRegDiffTSPred(Example1.TS, forward = 1L)
StatRegDiffTSPred(Example1.TS, forward = 2L)

# Predicting upon a times series with many NA values
data(Example2.TS)
StatRegDiffTSPred(Example2.TS, forward = 1L)

# On a matrix
Mat <- rbind(Example1.TS, Example2.TS)
StatRegDiffTSPred(Mat, forward = 1L)
}

