\name{TSPred-internal}
\alias{SSMpolynomial}
\alias{data.frame.na}
\title{Internal TSPred Functions}
\description{
 Internal TSPred functions
}
\usage{
SSMpolynomial(y, ord, H=0, Q=NA)
data.frame.na(..., row.names = NULL, check.rows = FALSE, check.names = TRUE,
    stringsAsFactors = FALSE)
}
\details{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).
}
\keyword{ internal }
