% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSPred-package.R
\docType{package}
\name{TSPred-package}
\alias{TSPred-package}
\title{Functions for Benchmarking Time Series Prediction}
\description{
Functions for time series pre(post)processing, decomposition, modelling, prediction and accuracy assessment. The generated models and its yielded prediction errors can be used for benchmarking other time series prediction methods and for creating a demand for the refinement of such methods. For this purpose, benchmark data from prediction competitions may be used.
}
\author{
Rebecca Pontes Salles

Maintainer: rebeccapsalles@acm.org
}
\keyword{package}
