% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArimaBaseTSPred.R
\docType{methods}
<<<<<<< HEAD:man/AutoArimaTSPred.Rd
\name{AutoArimaTSPred}
\alias{AutoArimaTSPred}
\alias{AutoArimaTSPred,vector-method}
\title{Method to predict according to an ARIMA model}
||||||| merged common ancestors
\name{AutoArimaTSPred}
\alias{AutoArimaTSPred}
\alias{AutoArimaTSPred,StQList-method}
\alias{AutoArimaTSPred,matrix-method}
\alias{AutoArimaTSPred,vector-method}
\title{Method to predict according to ARIMA model}
=======
\name{ArimaBaseTSPred}
\alias{ArimaBaseTSPred}
\alias{ArimaBaseTSPred,StQList-method}
\alias{ArimaBaseTSPred,matrix-method}
\alias{ArimaBaseTSPred,vector-method}
\title{Method to predict according to an ARIMA model}
>>>>>>> dca5fe21834e7ebe8b5c67fd5aae84b0980a48d6:man/ArimaBaseTSPred.Rd
\usage{
<<<<<<< HEAD:man/AutoArimaTSPred.Rd
AutoArimaTSPred(x, VarNames, forward = 2L, ...)
||||||| merged common ancestors
AutoArimaTSPred(x, VarNames, forward = 2L)
=======
ArimaBaseTSPred(x, VarNames, forward = 2L)
>>>>>>> dca5fe21834e7ebe8b5c67fd5aae84b0980a48d6:man/ArimaBaseTSPred.Rd

<<<<<<< HEAD:man/AutoArimaTSPred.Rd
\S4method{AutoArimaTSPred}{vector}(x, VarNames, forward = 2L)
||||||| merged common ancestors
\S4method{AutoArimaTSPred}{vector}(x, VarNames, forward = 2L)

\S4method{AutoArimaTSPred}{matrix}(x, VarNames, forward = 2L)

\S4method{AutoArimaTSPred}{StQList}(x, VarNames, forward = 2L)
=======
\S4method{ArimaBaseTSPred}{vector}(x, VarNames, forward = 2L)

\S4method{ArimaBaseTSPred}{matrix}(x, VarNames, forward = 2L)

\S4method{ArimaBaseTSPred}{StQList}(x, VarNames, forward = 2L)
>>>>>>> dca5fe21834e7ebe8b5c67fd5aae84b0980a48d6:man/ArimaBaseTSPred.Rd
}
\arguments{
\item{x}{object upon which the prediction will be made.}

\item{VarNames}{character vector with the variable names for which the prediction will be made;
by default it is NULL.}

\item{forward}{integer indicating the number of periods ahead when the prediction will be made;
by default it is 2L.}
}
\value{
It returns a list with components Pred and STD, containing the point prediction and the
estimated standard deviations, respectively. Depending on the class of the input parameter x, it
returns:

\itemize{
 \item For input class vector, it returns numeric vectors.
 \item For input class matrix, it returns matrices.
 \item For input class StQList, it returns list whose components are
  data.tables.
}
}
\description{
This method implements the predicted value and their standard deviation according to
<<<<<<< HEAD:man/AutoArimaTSPred.Rd
an ARIMA model.
||||||| merged common ancestors
ARIMA model.
=======
an ARIMA model as computed by the base package function arima().
>>>>>>> dca5fe21834e7ebe8b5c67fd5aae84b0980a48d6:man/ArimaBaseTSPred.Rd
}
\examples{

# Predicting one and two months ahead in time
data(Example1.TS)
ArimaBaseTSPred(Example1.TS, forward = 1L)
ArimaBaseTSPred(Example1.TS, forward = 2L)

# Predicting upon a times series with many NA values
data(Example2.TS)
ArimaBaseTSPred(Example2.TS, forward = 1L)

<<<<<<< HEAD:man/AutoArimaTSPred.Rd
||||||| merged common ancestors
# On a matrix
Mat <- rbind(Example1.TS, Example2.TS)
AutoArimaTSPred(Mat, forward = 1L)

=======
# On a matrix
Mat <- rbind(Example1.TS, Example2.TS)
ArimaBaseTSPred(Mat, forward = 1L)

>>>>>>> dca5fe21834e7ebe8b5c67fd5aae84b0980a48d6:man/ArimaBaseTSPred.Rd
\dontrun{
# With an object of class StQList
data(StQListExample)
VarNames <- c('ActivEcono_35._6._2.1.4._0', 'GeoLoc_35._6._2.1._1.2.5.')
<<<<<<< HEAD:man/AutoArimaTSPred.Rd
AutoArimaTSPred(StQListExample, VarNames)
||||||| merged common ancestors
AutoArimaTSPred(StQList_Example, VarNames)
=======
ArimaBaseTSPred(StQList_Example, VarNames)
>>>>>>> dca5fe21834e7ebe8b5c67fd5aae84b0980a48d6:man/ArimaBaseTSPred.Rd
}

}

