% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoArimaTSPred.R
\docType{methods}
\name{AutoArimaTSPred}
\alias{AutoArimaTSPred}
\alias{AutoArimaTSPred,vector-method}
\alias{AutoArimaTSPred,StQList-method}
\title{Method to predict according to an ARIMA model.}
\usage{
AutoArimaTSPred(x, VarNames, frequency = 12L, forward = 2L)

\S4method{AutoArimaTSPred}{vector}(x, VarNames, frequency = 12L,
  forward = 2L)

\S4method{AutoArimaTSPred}{StQList}(x, VarNames, frequency = 12L,
  forward = 2L)
}
\arguments{
\item{x}{object upon which the prediction will be made.}

\item{VarNames}{character vector with the variable names for which the prediction will be made;
by default it is NULL.}

\item{frequency}{integer indicating the frequency of time series; by default it is 12L.}

\item{forward}{integer indicating the number of periods ahead when the prediction will be made;
by default it is 2L.}
}
\value{
It returns a \code{data.table} with components Pred and STD, containing the point
prediction and the estimated standard deviations, respectively, for each variable.
}
\description{
This method implements the predicted value and their standard deviation according to
an ARIMA model.
}
\examples{

# Predicting one and two months ahead in time
data(Example1.TS)
AutoArimaTSPred(Example1.TS, forward = 1L)
AutoArimaTSPred(Example1.TS, forward = 2L)

# Predicting upon a times series with many NA values
data(Example2.TS)
AutoArimaTSPred(Example2.TS, forward = 1L)

\dontrun{
# With an object of class StQList
data(StQListExample)
VarNames <- c('ActivEcono_35._6._2.1.4._0', 'GeoLoc_35._6._2.1._1.2.5.')
AutoArimaTSPred(StQListExample, VarNames)
}

}
