% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedTSPred.R
\docType{methods}
\name{FixedTSPred}
\alias{FixedTSPred}
\alias{FixedTSPred,vector-method}
\alias{FixedTSPred,StQList-method}
\title{Method to predict a prefixed value for time series.}
\usage{
FixedTSPred(x, VarNames, value = 0L)

\S4method{FixedTSPred}{vector}(x, VarNames, value = 0L)

\S4method{FixedTSPred}{StQList}(x, VarNames, value = 0L)
}
\arguments{
\item{x}{object upon which the prediction will be made.}

\item{VarNames}{character vector with the variable names for which the prediction will be made;
by default it is NULL.}

\item{value}{numeric vector with the value for the prediction and the standard deviation.}
}
\value{
It returns a \code{data.table} with components Pred and STD, containing the point
prediction and the estimated standard deviations, respectively, for each variable.
}
\description{
This method implements the predicted value and their standard deviations as a
prefixed value.
}
\examples{

# Predicting one and two months ahead in time
data(Example1.TS)
FixedTSPred(Example1.TS, value = 0L)

# Predicting upon a times series with many NA values
data(Example2.TS)
FixedTSPred(Example2.TS, value = 0L)

\dontrun{
# With an object of class StQList
data(StQListExample)
VarNames <- c('Personal_07.')
FixedTSPred(StQListExample, VarNames, value = 0L)
}

}
