% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatRegDiffTSPred.R
\docType{methods}
\name{StatRegDiffTSPred}
\alias{StatRegDiffTSPred}
\alias{StatRegDiffTSPred,vector-method}
\alias{StatRegDiffTSPred,StQList-method}
\title{Method to predict according to the stational difference time series model.}
\usage{
StatRegDiffTSPred(x, StatDiff = 12L, forward = 2L, VarNames = NULL)

\S4method{StatRegDiffTSPred}{vector}(x, StatDiff = 12L, forward = 2L,
  VarNames = NULL)

\S4method{StatRegDiffTSPred}{StQList}(x, StatDiff = 12L, forward = 2L,
  VarNames = NULL)
}
\arguments{
\item{x}{object upon which the prediction will be made.}

\item{StatDiff}{stational differences of the time series; by default it is 12L.}

\item{forward}{integer indicating the number of periods ahead when the prediction will be made;
by default it is 2L.}

\item{VarNames}{character vector with the variable names for which the prediction will be made;
by default it is NULL.}
}
\value{
It returns a \code{data.table} with components Pred and STD, containing the point
prediction and the estimated standard deviations, respectively, for each variable.
}
\description{
This method implements the predicted value and their standard deviation according to
the regular difference time series model
\eqn{(1-B)^s y_{t}=a_{t}}{(1-B)^s y<sub>t</sub>=a<sub>t</sub>}.
}
\examples{

# Predicting one and two months ahead in time
data(Example1.TS)
StatRegDiffTSPred(Example1.TS, forward = 1L)
StatRegDiffTSPred(Example1.TS, forward = 2L)

# Predicting upon a times series with many NA values
data(Example2.TS)
StatRegDiffTSPred(Example2.TS, forward = 1L)

# On a matrix
Mat <- rbind(Example1.TS, Example2.TS)
StatRegDiffTSPred(Mat, forward = 1L)

\dontrun{
# With an object of class StQList
data(StQListExample)
VarNames <- c('ActivEcono_35._6._2.1.4._0', 'GeoLoc_35._6._2.1._1.2.5.')
StatRegDiffTSPred(StQListExample, StatDiff = 9L, VarNames = VarNames)
}

}
