\name{TSPred-internal}
\alias{SSMpolynomial}
\title{Internal TSPred Functions}
\description{
 Internal TSPred functions
}
\usage{SSMpolynomial(y, ord, H=0, Q=NA)
}
\details{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).
}
\keyword{ internal }
